+++ pkg/helmclient/delete.go
@@ -13,7 +13,10 @@ func (c *Client) DeleteRelease(ctx context.Context, namespace, releaseName strin
 	eventName := "delete_release"
 
 	t := prometheus.NewTimer(histogram.WithLabelValues(eventName))
-	defer t.ObserveDuration()
+	defer func() {
+		eventCounter.WithLabelValues(eventName).Inc()
+		t.ObserveDuration()
+	}()
 
 	err := c.deleteRelease(ctx, namespace, releaseName)
 	if err != nil {