+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added event count metrics for delete, install, rollback and update of Helm releases.
+
 ## [2.1.1] - 2020-09-21
 
 ### Security